refactor(vehicle_cmd_gate): move private headers (#1799)
